t9500: write "extensions.refstorage" into config
In t9500 we're writing a custom configuration that sets up gitweb. This
requires us manually ensure that the repository format is configured as
required, including both the repository format version and extensions.
With the introduction of the "extensions.refStorage" extension we need
to update the test to also write this new one.

Signed-off-by: Patrick Steinhardt <ps@pks.im>
Signed-off-by: Junio C Hamano <gitster@pobox.com>